Tests/LibWeb/Text/input/css/font-face-src-local-serialization.html

<script src="../include.js"></script>
<script>
    test(() => {
        let style = document.createElement("style");
        style.innerText = "@font-face { font-family: 'a1'; src: local('xyz'); }";
        document.head.appendChild(style);
        let sheet = style.sheet;
        println(sheet.cssRules[0].cssText);
    });
</script>
