Tests/LibWeb/Text/expected/css/font-face-src-local-serialization.txt

@font-face { font-family: "a1"; src: local("xyz"); unicode-range: "U+0-10ffff"; }
